requirements.txt
streamlit
pandas
numpy
requests
python-dotenv
google-generativeai
